+++ src/main/java/net/frozenblock/lib/recipe/mixin/ShapelessRecipeBuilderMixin.java
@@ -20,10 +20,11 @@ package net.frozenblock.lib.recipe.mixin;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import net.frozenblock.lib.recipe.api.ShapedRecipeBuilderExtension;
+import net.frozenblock.lib.recipe.api.ShapelessRecipeBuilderExtension;
 import net.minecraft.advancements.AdvancementHolder;
 import net.minecraft.core.component.DataComponentPatch;
 import net.minecraft.data.recipes.RecipeOutput;
-import net.minecraft.data.recipes.ShapedRecipeBuilder;
+import net.minecraft.data.recipes.ShapelessRecipeBuilder;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.crafting.Recipe;
 import net.minecraft.world.item.crafting.ShapedRecipe;
@@ -32,8 +33,8 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 
-@Mixin(ShapedRecipeBuilder.class)
-public class ShapedRecipeBuilderMixin implements ShapedRecipeBuilderExtension {
+@Mixin(ShapelessRecipeBuilder.class)
+public class ShapelessRecipeBuilderMixin implements ShapelessRecipeBuilderExtension {
 
 	@Unique
 	@Nullable
@@ -41,9 +42,9 @@ public class ShapedRecipeBuilderMixin implements ShapedRecipeBuilderExtension {
 
 	@Unique
 	@Override
-	public ShapedRecipeBuilder frozenLib$patch(@Nullable DataComponentPatch patch) {
+	public ShapelessRecipeBuilder frozenLib$patch(@Nullable DataComponentPatch patch) {
 		this.patch = patch;
-		return (ShapedRecipeBuilder) (Object) this;
+		return (ShapelessRecipeBuilder) (Object) this;
 	}
 
 	@WrapOperation(