+++ CHANGELOG.md
@@ -5,4 +5,6 @@ Make sure to clear this after each release
 Put changelog here:
 
 -----------------
+- Fixed custom Grass Color Modifiers to work 100% of the time.
+- Added `PlaceInAirBlockItem,` thanks to Qzimyion!
 - Updated to Kotlin 2.1