+++ CHANGELOG.md
@@ -8,3 +8,4 @@ Put changelog here:
 - Fixed custom Grass Color Modifiers to work 100% of the time.
 - Added `PlaceInAirBlockItem,` thanks to Qzimyion!
 - Updated to Kotlin 2.1
+- Added `ShapelessRecipeBuilderExtension`